+++ src/components/Card/style.module.css
@@ -1,0 +1,59 @@
+.Container {
+    margin: 10px;
+    padding-bottom: 30px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+    padding-top: 10px;
+}
+
+.Card {
+    border: solid 2px rgb(216, 213, 213);
+    border-radius: 15px;
+    background-color: rgb(237, 237, 237);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 20px 35px;
+    flex-direction: column;
+    margin: 20px;
+} 
+
+.Card:hover {
+    border: solid 2px rgb(131, 130, 130);
+    background-color: rgb(225, 220, 220);
+}
+
+.CardInfo {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    margin-top: 0;
+    margin-bottom: 30px;
+}
+
+
+.Button {
+    padding: 12px 18px;
+    border-radius: 30px ;
+    border: solid 1px rgb(56, 56, 56);
+}
+
+
+.Button:hover {
+    background: rgb(20, 115, 165);
+    cursor: pointer;
+    border: white;
+    color: white;
+}
+
+.PokeImg {
+    height: 140px;
+}
+
+.Error {
+    color: red;
+    font-size: 25px;
+}
