+++ src/components/InfoScreen/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const InfoScreen = () => {
+  return (
+    <h1>Esta es la pantalla donde aparecerá la info del Pokemon</h1>
+  )
+}
+
+export default InfoScreen
